chore(components): integrate existing header and footer as separate components
