+++ venue-reservation/src/app/components/layout/Footer.tsx
@@ -1,0 +1,21 @@
+"use client";
+import React from 'react';
+
+const Footer = () => {
+    return (
+      <footer className="bg-gray-800 text-white py-8">
+        <div className="container mx-auto px-4 text-center">
+          <p>&copy; 2024 MyVenue. All Rights Reserved.</p>
+          <nav className="mt-4">
+            <ul className="flex justify-center space-x-4">
+              <li><a href="#" className="hover:underline">Privacy Policy</a></li>
+              <li><a href="#" className="hover:underline">Terms of Service</a></li>
+              <li><a href="#" className="hover:underline">Contact Us</a></li>
+            </ul>
+          </nav>
+        </div>
+      </footer>
+    );
+  };
+
+  export default Footer;
